Invert y-axes when making zonal mean plots in benchmark.py
The default axis settings plotted vertical coordinate 0 at the bottom
of the plot and 1 at the top of the plot.  But because the hybrid
coordinate goes from 1 at the surface to 0 at the top of atmosphere,
the zonal means were inverted.

We put a call into ax.invert_yaxis() to flip the vertical axis.

TO DO: Need to define the vertical coordinate from the metadata in
the netCDF file.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>