Use max of abs 5th and 95th percentiles to set symmetric restricted range
The restricted range is the current placeholder in GC Classic benchmark
plotting routines. Existing GAMAP benchmark fixed range is hard-coded in
GAMAP. The advantage of using 5% and 95% percentiles is the removal of
outliers, although this is only true in one direction (positive or
negative) if using a symmetric colorbar range. But a symmetric colorbar
is necessary to center colormap white on zero.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>