Add function find_varnames to core.py
In module gcpy/gcpy/core.py, we have now added the function
find_varnames.  This can be used to return a list of variable names
from an xarray Dataset object.  You can either return all variables
in the Dataset, or get a subset of those variables that match
a certain text pattern.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>