Remove IJ_SOA_S_ from bpch names conversion
There was a recent update to GEOS-Chem Classic that changed the binary
aerosol mass diagnostic names in such a way that now causes an error
when converting binary names. This should be revisited in the future.
The quick fix for now is to remove the line from the conversion name
dictionary.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>