Add script for generating 1-month benchmark plots
This script can be run from the command line as:
$ python run_1mo_benchmark.py

To use this, you should copy it to where you want to generate benchmark
output and then edit the paths and filenames in the file. Also edit
the ref and dev strings that will be inserted into plot titles. Archive
the run script when archiving a benchmark.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>