Remove ONIT and PMNO2 from lumped species lists; change BrO2 to BRO2
ONIT and PMNO2 are defined in the KPP mechanism (.eqn) file, but are not
actually used in any reactions. KPP thus ignores them and doesn't actually
include them as species in the mechanism.

Inclusion of BrO2 was a typo. It is not a bromine species, but the product
of BENZ + RO2.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>