Add bug fix and minor edits to run_1mo_benchmark.py
Needed to add "benchmark." before "make_gcc_1mo_benchmark_emis_tables".

Also rearranged the settings at the top of the file for easier editing.
Users should generally only need to change the path to their run directories,
the strings for ref and dev versions, and the date (YYYYMMDD) of the
simulation. The name of spcfile, hcofile, etc. shouldn't change often so
we can use templates and simply fill in the date.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>