Add functionality to make emissions plot file per benchmark species category
The files are saved in the same directory structure as the benchmark
concentration plots, with file name {category}_Emissions.pdf. The
bookmarks are nested, with each species serving as top-level bookmark and
{species}_{HEMCO category} serving as lower-level bookmark.

Plotting emissions by benchmark category and by HEMCO category are both
on by default in the 1-month benchmark script.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>