Add GCHP vs GCC and GCHP vs GCHP to benchmark run script
Because of the addition of GCHP there are extensive changes to
previous variable names. The GCC vs GCC comparison should still
function the same as previously. Logicals at the top of the file
now toggle what comparisons to run, and what materials (e.g conc plots,
emissions tables) to generate.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>