Add fixes for new emissions inventory diagnostic names to make_benchmark_emis_tables
Emissions inventory diagnostic names are now in the format InvNAME_SPC
instead of InvSPC_NAME. The new emission_inventories.json lists the
inventory names used in GC 12.4.0. This change in name formatting also
required modifications in create_total_emissions_table so that we obtain
species name and it's properties correctly.

Also added a fix to compare_varnames in core.py. Refdata is used to identify
a variable as 1D and 2D, and it is now also used for 3D fields instead of
devdata.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>