Add arguments to pass the variable name for surface area for emis tables
We have added the ref_area_varname and dev_area_varname variables to
the routine make_benchmark_emis_table.  This allows us to specify
the variable that holds grid box surface areas in the dev and ref
datasets.  For GCC, this variable is usually called AREA and for
GCHP this is Met_AREAM2.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>