Fix for run_1mo_benchmark.py script for non X-window environments
Long story short: Some versions of matplotlib (perhaps prior to v3.0)
try to always attach to an X server.  This can cause issues if you
are running in a computational queue with no interactive display
capabilities.

We now set an environment variable at the top of the benchmark driver
script run_1mo_benchmark.py to explicitly tell matplotlib not to try
to connect to an X server.  We have also added comments to the docstring
at the top of the script.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>