Rename WL1 to 550nm in make_benchmark_aod_plots
For GCHP, the diagnostic exports are populated before input.geos is
read, and such, the AOD diagnostic names have "WL1" in them.  But
in GCC, this will get replaced to 550nm.

To facilitate a direct comparison between GCHP and GCC, we have
put code into benchmark.make_benchmark_aod_plots to replace the text
WL1 with 550nm in all Dataset variables.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>