Check if emissions variables have the same dims before plotting
In method benchmark.make_benchmark_emis_plots, we have added a check
to make sure that all variables in Dev and Ref have the same dimensions
before plotting.  For example, in GC 12.3.0 some emissions variables
were 2D but in 12.3.2 they were updated to be 3D.  This would cause
an error when trying to take the column sum.  Variables with mismatching
dimensions will now be skipped and a warning message will be printed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>