Update GCHP diff of diffs code in run script to work with any MAPL cs grid
The dimensions and dimension names changed in MAPL version 1.0.0 for the
output cubed sphere grid. Prior to v1.0.0 horizontal dimensions are
called lat and lon. Starting v1.0.0 there are three horizontal dimensions:
Xdim, Ydim, and nf. Both grids work with the plotting functions in
the gcpy benchmark code. However, doing xarray math functions do not
work when operating on DataSets or DataArrays with differing grids.
This is only a problem in gcpy code within the calculation of difference
of differences in GCHP in the 1-month benchmark plot generation script.

This update detects if the ref and dev GCHP datasets are differing grids.
If they are, then ref is converted to the new MAPL grid which is assumed
to be the same as in dev. The differences are then calculated and
saved to a netcdf file.

Pre-existing code to copy attributes is removed since it is not
necessary with the xr.set_options keep_attrs=True option.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>