Initial commit of log scale color updates
Modify the routines in benchmark.py to plot the top row with a log
color scale if the log_color_scale=True keyword is passed to e.g.
compare_single_level and compare_zonal_mean.

NOTE: There seems to be an issue that causes the plotting to halt
with an error in pdf.savefig() if the Ref and Dev data are both all
zeroes.  We are currently investigating.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>