Do not take log (with LogNorm or LogFormatter) if data is all zeroes
We have updated the logic in compare_single_level and compare_zonal_mean
so that we will not call mpl.colors.LogNorm or mpl.ticker.LogFormatter
if all elements of the Ref or Dev data arrays are zero.  This will
prevent numerical errors caused by taking the log of zero.

Also note: if all of the data is zero in any of the plot windows, then
we will only display a single tick mark located at zero.  This is
seems more intuitive and will be easier to tell at a glance that there
is only one value in the entire plot window.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>