Improve the test for all zeroes in compare_zonal_mean
In compare_zonal_mean, we now use the criterion np.all(zm_diff == 0)
to determine if we should manually set one colorbar tickbar at 0.0,
for subplots (2,0) and (2,1).

The prior test used zm_fracdiff... but zm_fracdiff can have NaN values
due to div/by zero if both Ref and Dev data arrays are zero.  So this
test is more robust.  It also matches what we do in compute_single_level.

NOTE: This test is only used for colorbar tickmark placement.  It is
not used for computation.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>