Return list of significant differences from level and zonal plots
compare_single_level and compare_zonal_mean now both return a list
of species (or quantities) that have a fractional difference of
greater than 0.1 = 10% anywhere on the grid.  This will be used for
helping to prepare the benchmark form.

Further work is necessary in order to write out a file with all of
this information.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>