Updates to enable generating benchmarks emissions tables with new MAPL grid
The new MAPL grid output does not include lat and lon as dimensions,
making parts of the emissions tables code result in error when using
GCHP files starting in 12.5.0. Previously used xarray open_mfdataset
also resulted in an error in the new MAPL grid. The code used for
generating emissions tables is slightly altered to avoid use of
open_mfdataset and to remove the dependency on dimension names. The
area is now appended to the GCHP DataSet with the same name as used
in GEOS-Chem Classic to make grid type checking unnecessary. This also
enables comparing against lat-lon output of GCHP, which is possible in
12.5.0.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>