Add minor changes in run_1mo_benchmark.py for updated run directory structure
Starting in GC 12.5.0, output files for both GCC and GCHP are put in
OutputDir and plots will be put in the Plots dir. This is now fixed in the
file paths.

Also removed GCC and GCHP from plot titles. The version run directories (and
strings) are typically named with GC_x.y.z and GCHP_x.y.z so there's no need
to include an extra GCC or GCHP in front.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>