Set units of column AOD variables to "1" in benchmark.py
In routine make_benchmark_aod_plots (in benchmark.py), there can be
an assertion error if the units of the Ref and Dev AOD variables
are different.  For example, if one uses "unitless' (such as from
bpch output) and the other uses "1", then this can cause the
plotting to halt with an assertion error.  This is because the
default is to keep the same units as the Ref and Dev datasets.

The simple fix is to manually set the units of the column AOD variables
to "1" so that they will be the same, regardless of what the units
of "Ref" and "Dev" are.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>