Add error handling for CS to CS comparison at different resolutions
CS to CS regridding is not implemented in gcpy. Ref and dev at different
CS resolutions can only be compared on a lat-lon grid. This is an open
issue, but now there is error handling to exit out with a helpful message
if attempting to regrid CS to CS due to the comparison grid.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>