Emissions tables will now print NaN for missing variables
If an emissions (or inventory) variable is present in the Ref dataset
but not in the Dev dataset (or vice-versa), we will now print its
total as "nan" (not-a-number).

This update was done to reduce confusion for members of the GCSC who
view the benchmark results.  Before this commit, only emissions
variables that were present in both Ref and Dev datasets had totals
printed.  The nan values indicate emissions variables that are either
introduced to or retired from a given benchmark simulation.

We will make similar modifications for the plotting routines in
subsequent commits.

Also note: core.compare_varnames now also returns the refonly and
devonly variables, which are lists of variables that are only present
in Ref or only in Dev.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>