compare_varnames now returns 2D or 3D vars in refonly, devonly lists
The refonly and devonly arguments returned by core.compare_varnames
now only contain plottable variables (i.e. those that are not
also included in commonvarsOther).  This will facilitate combining
these into variable lists so that we can plot null results for
species that are in Ref but not in Dev (or vice-versa).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>