Update MDA8 ozone timeseries example script
The previous example scripts worked with external datasets and bpch input.
The mad8_o3_timeseries.py script added here uses hourly netCDF output from
GEOS-Chem. The resulting dataset is saved to a new netCDF file and plotted.

This script also fixes an issue in the resample method. The usage of this
method in the other scripts was for older versions and no longer worked.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>