Now use shallow copy when setting cmap2 in compute_single_level
When defining the cmap2 color map for Ref and Dev (i.e. the first row),
we must use a shallow copy operation when defining cmap2.  Otherwise,
this will cause the Ref and Dev plots to be plotted with gray
striping.

I believe this is because the cmap2.set_bad() property also causes
cmap1 to have the bad colors set to gray.  In other words, change:

  cmap1 = WhGrYlRd
  cmap2 = WhGrYlRd
  cmap2.set_bad(color='gray')

to
  cmap1 = WhGrYlRd
  cmap2 = copy.copy(WhGrYlRd)
  cmap2.set_bad(color='gray')

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>