Fixed a typo in compare_single_level when use_cmap_RdBu
In routine compare_single_level (in benchmark.py), there was a typo
in the definition of absmax when use_cmap_RdBu=True but when Ref
had all NaN values.  In this case, the values of vmin_ref and
vmax_ref were used to define absmax, but we should have used
vmin_dev and vmax_dev.  Now fixed.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>