Fix striping for fractional difference plots on cubed-sphere grids
In routine compute_single_level, we have moved the longitude masking
of the fracdiff array (between  178 and 182 degrees, which may be
necessary to avoid a spurious boundary line) to after the operation
where we convert Inf to NaN.  This seemed to have an adverse affect
on the call to pcolormesh.

NOTE: The longitude masking of fracdiff is only done for cubed-sphere
grids.  We have confirmed that this update does not affect plotting
frac. diff. for lat-lon grids.

This commit should resolve gcpy Github issue #15.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>