Initial modifications for more efficient masking for global totals
We are working on a more efficient (and faster) method for
creating a mask so that we can print tropospheric mass totals
(in addition to global mass totals).  We have made several
modifications in both benchmark.py and core.py.

NOTE: For now, the tropospheric mass option does not work, but we
will stash these updates here for the time being.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>