Enable creation both whole-atmosphere and tropospheric mass tables
Added several updates in benchmark.py and core.py to enable printing
of species masses for the entire atmosphere and only for the
troposphere.  The key updates are:

In core.py:
(1) Moved routine "normalize_colors" here from "benchmark.py", as this
    can be used by other plotting codes than just the benchmark routines.
(2) Added convenience routine get_area_from_dataset, which returns
    the proper area variable (either AREA or Met_AREAM2) from
    a given dataset.
(3) Simplified the code in get_dataarray_shape.
(4) Added convenience routine get_variables_from_dataset, which
    will return a list of DataArray objects from a dataset.
    This makes it easier to extract several variables from
    a Dataset at once.

In benchmark.py:
(1) Now create tropospheric mask arrays for Ref and Dev datasets
    in make_benchmark_mass_tables.  We have to make separate masks
    in case Ref is from GCC and Dev is from GCHP, etc.
(2) We now pass meterological variables and mask variables separately
    (via a dictionary named "met_and_masks") from routine
    make_benchmark_mass_tables to routine create_global_mass_tables.
(3) The application of the tropospheric mask arrays is now done
    for each species in routine print_totals.  If the dictionary
    "met_and_masks" is passed to print_totals, then it will print
    the tropospheric-only totals.  Otherwise, it will print totals
    from the whole atmosphere.

NOTE: The tropospheric masking is only applied to global mass tables
and not to the emissions total tables.  We have verified that the
code produces identical results for the emissions tables as prior to
this commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>