Function get_variables_from_dataset now returns an xarray Dataset
The convenience function get_variables_from_dataset now returns
an xarray Dataset object instead of a list of DataArrays.
This is intended to be used when you need to extract several
variables from a datalist.  If not all the variables are found,
an error will be raised (such as we do in the global mass tables
code.)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>