Now exclude "anchor" from open_dataset and open_mfdataset calls
According to xarray issues:
   https://github.com/pydata/xarray/issues/3286
   https://github.com/pydata/xarray/issues/1378

The open_mfdataset function has problems in creating a merged
dataset from multiple files in which variables have repeated
dimension names.  The easiest thing to do in this case is to
prevent such variables from being read in.

We now have added the drop_variables keyword to avoid reading
in the "anchor" variable in all calls to open_dataset and
open_mfdataset in both benchmark.py and core.py.  This variable is
only present in GCHP-created netCDF files using MAPL v1.0.0, which
is in GCHP 12.5.0 and later.

This commit should resolve GCPy issue #26:
   https://github.com/geoschem/gcpy/issues/26

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>