Fix bug in benchmark plotting when expliclty passing cmpres as cubed sphere
The comparison resolution passed to compare_single_level is supposed to
be able to be either cubed sphere or lat-lon. However, if it was cubed
sphere then the function would fail due to a type error. This is because
cmpres is passed as string but the function expects it to be integer if
cubed sphere. All cubed sphere comparisons to date have not passed cmpres
explicitly, and hence this bug has not been previously detected. The
quick fix is to cast cmpres to integer if cubed sphere and passed
explicitly.

A related issue occurred in compare_zonal_mean although cmpres should
never be passed as cubed sphere for that function. There is now handling
to override the passed cmpres value to be '1x1.25' if it is passed as
cubed sphere. A warning is printed indicating this override is happening.


Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>