Fix logic when defining masks to compute tropospheric-only masses
According to the Numpy manual:
   https://docs.scipy.org/doc/numpy/reference/maskedarray.generic.html

"When an element of the mask is False, the corresponding element of
 the associated array is valid and is said to be unmasked. When an
 element of the mask is True, the corresponding element of the
 associated array is said to be masked (invalid)."

The algorithm that we had been using defined masks that actually
produced the opposite of the desired result (i.e. True everywhere
up until the tropopause, and False above it).  We have now inverted
the logic (i.e. the mask values in each column are is False up to
tropopause location, and True above that).  This now produces the
proper results.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>