Now use a more robust algorithm for add_missing_variables
In order to avoid an error such as:
   ValueError: coordinate ilev has dimensions ('ilev',), but
   these are not a subset of the DataArray dimensions
   {'time': 1, 'lev': 72, 'lat': 46, 'lon': 72}

We have rewritten the algorithm in benchmark.add_missing_variables
so that we do the following:

(1) Make a shallow copy (i.e. copy.copy()) each DataArray that is
    in Ref but not in Dev.
(2) Reset the values in the copied DataArray to NaN everywhere.
(3) Merge the DataArray into Dev.
(4) Follow steps (1) to (3) for each DataArray that is in Dev but
    not in Ref.

Also: in routine make_global_emissions_table, we follow a similar
strategy where we make a shallow copy of a DataArray variable and
set it to NaN.

The shallow copy -- copy.copy() -- is necessary in order to avoid
having the original DataArray be affected when we reset the values
of the copied DataArray to NaN.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>