Fix some plotting issues that may be related to matplotlib 3
In core.py we now test if matplotlib is version 3 or higher,
and then normalize the data range [-1,1] if the min and max of
the data are both zero or both NaN.

We have also added a new function core.one_cb_tick_in_center(),
which returns the proper value to place one tick mark in the
center of the colorbar for all zero or all undefined data.
This value is [0.0] for matplotlib3 and [0.5] for older versions.

Also added some modifications to benchmark.py to make it easier to
abstract the plotting code into routines that can be called
multiple times.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>