Further fixes to handle matplotlib 3.1.0 and higher
We have now made a robust fix to so that a single colorbar tickmark
will be placed at the center of the colorbar (with the appropriate)
label if the data is zero everywhere or undefined everywhere.  A
change introduced in matplotlib 3.1.0 and higher caused the existing
algorithm not to function properly for the Ref and Dev plots.

Also made a few minor fixes to benhcmark routines.  In particular,
you can now use './' to define the destination (dst) variable
for the emissions tables.  Also removed a useless error check that
was causing GCHP vs. GCC emissions tables not to be printed out.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>