Fix lingering issues with tickmark placement
Realized that the issue with shifted colorbar tickmarks is not due
to matplotlib 3.1.1 vs. earlier versions.  The issue is that if we
are using the RdBu colormap (as we do for diff-of-diffs plots),
then we must use cb.set_ticks([0.0]), which places the tick
in the middle of the colorbar.  If we are not using the RdBu
colormap, then we msut use cb.set_ticks([0.5])

Other updates in this commit:
(1) Removed core.one_cb_tick_in_center().

(2) Added print statements to better separate lines of text that
    get printed when we create PDF files.

(3) In benchmark/run_1mo_benchmark.py, we have removed the if plot_conc:
    block in the diff-of-diff section.  If we had plot_conc=False,
    then this would not print anything when diff-of-diff plots were
    selected.  At present the concentration plots are the only option
    in diff-of-diffs, but if we add others, we can restore the
    if plot_conc: block.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>