Initial modifications for 1-year benchmark outputs
We have made several modifications to facilitate creating of the
1-year benchmark plots.  For now we will focus on the GCC vs GCC
comparisons, and add the others later.

We now have to pass the "interval" argument as a vector.  This contains
the number of seconds of each month.

We have also updated get_gcc_filepath and get_gchp_filepath to return
a list of filepaths for a given year, or a single file path.

Added the benchmarks/run_1yr_benchmark.py script as the benchmark
driver.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>