Add function get_troposphere_mask; other minor updates
We have now added the function get_troposphere_mask, which can return
a mask of the tropospheric grid boxes given a Dataset object that
contains Met_BXHEIGHT and Met_TropLev.  This is used to replace
hard-coding in make_benchmark_mass_tables and make_benchmark_oh_metrics.
This routine also works if the time dimension is greater than 1, as
is true for 1-year benchmarks.

Also make sure that the interval keyword in various routines
used a default value of [2678400.0], which is the number of seconds
in July (our 1-month benchmarking month).  We have placed this in
a list so that the time loop in convert_units will work.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>