Update core.get_filepaths to accept a list of collection names
The core.get_filepaths routine only returned file paths for a single
collection.  We have now updated it so that it accepts a list of str
(or a single str) containing the names of each collection.  This will
avoid having to call core.get_filepaths multiple times for multiple
collections, and then having to concatenate the output in the
calling routine.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>