Update the aws.s3_download_cmds_from_log function
We have made some changes to the aws.s3_download_cmds_from_log
command.  This can generate a script that can be used to download
data from the GEOS-Chem S3 bucket to an EBS instance (or elsewhere).

(1) Added a force_download keyword.  Setting this to True will always
    download files, even if they are already present on your EBS
    volume or local disk.  (This would be appropriate for a first-time
    copy.)  Setting force_download to False will only copy those files
    that are not found on your EBS volume or local disk.

(2) In the HEMCO/GMI/v2015-02 folder, we create files for IPMN and NPMN
    from PMN and RIPA, RIPB, RIPD from RIP.  We now copy and do not use
    symbolic links.

(3) Update the examples/aws-cloud/create_s3_download_script.py to
    give the user the option to set force_download to True or False.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>