Added examples/dry-run/parse_dryrun_output.py script
This is an example script that can be used to parse GEOS-Chem or
HEMCO dry-run log files and create a list of unique files as well
as data download scripts.

At present, the script will only create a download script from
the AWS s3://gcgrid bucket.  More to follow soon.

Also note: this script is independent of GCPy modules, and only
relies on the "os" and "sys" packages of core Python.  Therefore,
it can also be distributed independent of GCPy, such as in GEOS-Chem
or HEMCO run-directories.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>