Add TransportTracersBenchmark code to gcpy/budget_tt.py
We have modified the prior transporttracers_budget.py script to make
it more general, and so that it can be called directly from the
run_1yr_benchmark.py driver script.  This is now included in
gcpy/budget_tt.py.

We can also copy this into a similar module for the AOD budgets,
which will be done in a subsequent commit.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>