Update budget_aer.py to create aerosol burdens and budgets
We have updated the budget_aer.py module so that it now correctly
computes annual average aerosol burdens and global mean AOD's.

NOTE: We still have to add the computation of global mean AOD for
dust.  This is because we have to sum the AOD diagnostics to obtain
an overall dust AOD.  That will be done in subsequent commits.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>