Activated mass tables in 1-year benchmark output (FullChem for now)
We now call make_benchmark_mass_tables for the 1-year FullChemBenchmark
from run_1yr_benchmark.py.  This will print seasonal masses of
species.  Tables are saved to the Tables subdirectory (so the path
is typically devdir/Plots/Tables, where devdir is the run directory
of the Dev version.)

Also updated the changelog accordingly.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>