Add module to compute mean OH from log files
Added gcpy/mean_oh_from_logs.py, which computes the mean OH from
1-year FullChemBenchmark simulations.

NOTE: We need to doublecheck the existing make_benchmark_oh_metrics
routine because that seems to be inaccurate.  It may be due to the
fact that we use monthly mean inputs.  For now the mean OH from the
log files is much more useful.

Also updated the changelog.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>