Add fixes for getting GCHP restart file and StateMet_inst fields from different directories
The GCHP restart file is stored in the top-level run directory, while the
StateMet_inst file is stored in OutputDir. Code in run_1mo_benchmark.py has
been updated to get the files in the two different locations instead of
assuming they are in the same directory.

Also fixed a few other typos in comments and print statements.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>