Update 1-month benchmark script to test gcpy without modification
To test gcpy, copy the script anywhere you want to run the test and set
gcpy_test to True at the top. Edit options to hone in on what you want
to test. Benchmark artifacts will be created locally in new folder
called Plots. If using outside of Harvard, edit the maindir to
point to your test data.

Also in this update, skip the GCHP budget table since not many budget
diagnostics are configured for GCHP benchmark. Also adjust the budget
table prints in the script to reflect that they are not comparisons.
Turn off GCHP OH metrics as well. Add prints in the script to explicitly
state what is being skipped and why.