Use scientific notation for transport tracer operations budget tables
Using six decimal places is not enough precision for some of the passive
species in the operations budget tables. For the 1-year transport tracer
benchmarks, we now pass the pd_float_format argument to specify usage of
scientific notation.

Also fixed a minor typo in the call to make_operations_budget_table to
properly pass the label to that routine.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>