Use version strings not subtitles in 1-mo benchmark sig diff output files
In formal 1-month benchmarks the ref and dev version strings are the same
as the ref and dev strings used in plot subtitles. However, reusing the
1-month benchmark script for informal model run comparisons may involve
changing the subtitles to be more descriptive, and include characters
such as spaces and commas. This update prevents using those within the
sig diff filenames.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>