Now use Load=yaml.FullLoader in yaml.load to avoid warnings
We have changed YAML file open statements from:

   yaml_load_file(open(filename))

to

   yaml.load(open(filename), Loader=yaml.FullLoader)

in order to avoid a bunch of warnings about how opening a YAML file
without a Loader specified is unsafe.

This update was tested with a 1-month GCC vs. GCC test benchmark output.

NOTE: The species_database.yml file was NOT updated.  Only the commands
to read it in from disk.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>