Update convert_units to handle multiple time increments
The previous update to convert_units enabled use with datasets without the
time dimension, such as a GCHP restart file. This temporarily removed
handling for multiple times, such as is needed for 1-year benchmarks.
Convert_units can now handle both cases and with throw an error if the
dataset does not have a time dimension but the number of time increments
passed to the function is greater than one.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>