Enable initial and final mass in GCHP radionuclide budget tables
Previously the initial and final mass, plus the accumulation calculated
from them, was disabled for GCHP. Calculating these is now possible with
recent updates to the GCHP restart file for transport tracers
in 12.8.1. This update should not be used with GCHP files prior to
12.8.1 unless AREA is manually added to the initial and final restarts
prior to running. AREA may be extracted from the output restart file
from any run with 12.8.1 and later.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>