Fix bugs calculating lumped species with some or all missing constituents
This bug manifested when either none of the constituent species defined
were found in the dataset or when the first constituent species was not
found. The issue was improper initialization of the new lumped species
dataarray. Proper initialization only occurred in the first iteration of
the loop over constituent species listed in the lumped species yaml file.
If that species was missing from the dataset then the new lumped species
would be initialized to the last constituent species of the last lumped
species calculated.

Also in this commit: use verbose True by default in function
add_lumped_species_to_dataset and add docstring to that function.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>