Rework benchmark function check_units and add data_unit_is_mol_per_mol
Prior to this commit check_units served to both check that ref and dev
units were the same, as well as convert ref and dev data to ppb if
mol/mol. Now check_units only checks if ref and dev units match, returning
a boolean. New function data_unit_is_mol_per_mol checks if an input
dataset has units of mol/mol, using a list of acceptable mol/mol formats
known to be used in GEOS-Chem files, both netcdf and bpch. The conversion
to ppb is now directly done in the benchmark comparison plot functions to
be explicit about the value change.

This commit also removes an extraneous call to check_units.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>