Use species molecular weights to convert to ug/m3 in benchmark conc plots
By default the Aerosols and Secondary_Organic_Aerosols benchmark
category species are now converted to ug/m3 in benchmark concentration
plots. Molecular weights are retrieved from the species database with
the exception of Complex_SOA and Simple_SOA which are lumped. The
molecular weights for these are hard-coded where the conversion to ug/m3
is done in the plotting code. Eventually the conversion should be
abstracted to a convert units function, and molecular weights for use
in lumped species should be stored elsewhere for retrieval.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>