Use dir name not version string for species db path in 1-mon benchmark run
It is possible to use different version strings in benchmark runs from
directory names. However, the version string rather than the directory
name was used to set the path to the species database file used for
generating the 1-month benchmark. This causes an error if the version
string does not match the directory name.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>