gcpy/oh_metrics.py now computes OH metrics from the Metrics collection
Added the gcpy/oh_metrics.py script (which is based on the metrics.py
script that gets added to each fullchem run directory) for printing
the mean OH, MCF lifetime, and CH4 lifetime from the Ref and Dev
version.  This now uses the new Metrics collection in 13.0.0 and
later GEOS-Chem versions instead of relying on output to log files.

NOTE: For the time being, we will keep gcpy/mean_oh_from_logs.py,
as this is used with benchmarks prior to 13.0.0.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>