Add bug fix for computing seconds in benchmark month for GCHP
In run_1mo_benchmark.py, gchp_*_sec_in_bmk_month needs to be multiplied
by two to get the number of seconds in the benchmark month. This is needed
because the gchp_ref_date and gchp_state_date are for the middle of the
month, not the actual start date of the month. Without this fix, emission
tables and operation budget tables for GCHP will not be generated.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>