Only archive benchmark species and lumped species YAML files if missing
In order to prevent re-copying of the benchmark species and lumped
species files to disk, we have added an if block that only does the
copy operation if the archived file is not present on disk.  This is
done in gcpy/util.py in methods archive_benchmark_species and
archive_lumped_species.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>