Set verbose to False by default in add_lumped_species_to_dataset
The output printed by add_lumped_species_to_dataset is excessive and should
only be included if the verbose option is passed. For some reason, verbose
was set to True by default in this function. It is now set to False by
default.

The formatting of other print statements has also been cleaned up here.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>