Only copy run_1yr_benchmark.py to benchmark folder if it exists
The "if not exists(dest): copyfile(curfile, dest)" block will attempt
to copy the benchmark script to a non-existent directory. This error
was encountered when trying to plot the benchmark output for only
gcc_vs_gcc but none of the other options.

We need to change this to "if exists(dest): copyfile(curfile, dest").

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>