Fix bug in regrid_comparison_data where overloaded operator undefined
One usage of the += operator in function regrid_comparison_data caused GCPy
to crash when regridding GCHP for comparison with GC-Classic data in
benchmarks. The case it was being used for was not defined in the dask
core package. All other usages of += are simple integers so no problems.
The fix is to change the equation format from A += B to A = A + B which
was presumably the intent.